requirements.txt
numpy==1.26.4
pandas==2.2.2
opencv-python==4.10.0.84
scikit-learn==1.5.1
tensorflow==2.17.0
keras==3.4.1
imutils==0.5.4
matplotlib==3.9.0
Pillow==10.4.0
joblib==1.4.2
flask==3.0.3
gunicorn==22.0.0
